requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
algoliasearch==1.17.0
apscheduler==3.3.1
argon2-cffi==16.2.0
beautifulsoup4==4.6.0
cairocffi==0.7.2          # via cairosvg, weasyprint
CairoSVG==1.0.22          # via weasyprint
certifi==2018.10.15       # via requests
cffi==1.7.0               # via argon2-cffi, cairocffi, weasyprint
chardet==3.0.4            # via requests
click==6.6                # via rq
coinbase==2.1.0
configparser==3.5.0       # via flake8
contextlib2==0.5.5        # via raven
coreapi==2.3.0            # via django-rest-swagger, openapi-codec
coreschema==0.0.4         # via coreapi
cssselect2==0.2.1         # via weasyprint
cssselect==0.9.2          # via premailer
cssutils==1.0.1           # via premailer
django-activity-stream==0.6.2
django-allauth==0.31.0
django-braces==1.11.0     # via django-oauth-toolkit
django-cors-headers==1.2.2
django-countries==4.2
django-filter==1.1.0
django-hashers-passlib==0.3
django-oauth-toolkit==0.12.0
django-redis==4.4.3
django-rest-auth==0.9.1
django-rest-swagger==2.1.2
django-rq==0.9.5
django-tagulous==0.12.0
Django==1.11.19           # via django-allauth, django-hashers-passlib, django-oauth-toolkit, django-rest-auth, django-rq
djangorestframework-jwt==1.8.0
djangorestframework==3.6.2
docopt==0.4.0             # via mandrill
dry-rest-permissions==0.1.6
enum34==1.1.6             # via argon2-cffi, flake8
exactonline==0.3.2
flake8==3.5.0
funcsigs==1.0.2           # via apscheduler
futures==3.0.5            # via apscheduler, isort
gunicorn==19.6.0
honcho==1.0.1
html5lib==0.999999999     # via weasyprint
idna==2.7                 # via requests
isort==4.3.4
itypes==1.1.0             # via coreapi
jinja2==2.9.5             # via coreschema
lxml==3.6.4               # via premailer
mailchimp3==2.0.7
mandrill==1.0.57
MarkupSafe==1.0           # via jinja2
mccabe==0.6.1             # via flake8
mysqlclient==1.3.7
oauthlib==2.0.1
openapi-codec==1.3.1      # via django-rest-swagger
passlib==1.6.5            # via django-hashers-passlib
pdfrw==0.4                # via weasyprint
Pillow==3.3.0
premailer==3.0.1
pycodestyle==2.3.1        # via flake8
pycparser==2.14           # via cffi
pycryptodome==3.7.0
pyflakes==1.6.0           # via flake8
PyJWT==1.4.0              # via djangorestframework-jwt
Pyphen==0.9.4             # via weasyprint
python-dateutil==2.5.3
python-openid==2.2.5      # via django-allauth
pytz==2016.10             # via apscheduler, django, tzlocal
raven==6.0.0
redis==2.10.5             # via django-redis, rq
requests-oauthlib==0.8.0  # via django-allauth
requests==2.20.0
rest-framework-generic-relations==1.1.0
rq==0.6.0                 # via django-rq
simplejson==3.10.0        # via django-rest-swagger
six==1.10.0
slackdown==0.0.3
slacker==0.9.30
stripe==1.56.0
tinycss2==0.6.1           # via cssselect2, weasyprint
tzlocal==1.3              # via apscheduler
uritemplate==3.0.0        # via coreapi
urllib3==1.24             # via requests
WeasyPrint==0.42.3
webencodings==0.5         # via html5lib, tinycss2
XlsxWriter==1.0.4

# The following packages are commented out because they are
# considered to be unsafe in a requirements file:
# setuptools                # via apscheduler, html5lib

gspread==3.1.0
oauth2client==4.1.3
